refactor: extract orphaned repository detection to custom hook
Extract duplicate logic for detecting orphaned GitHub vector store repositories into a reusable custom hook `useGitHubVectorStoreStatus`. This improves code maintainability and ensures consistent behavior across components.

- Create new hook `useGitHubVectorStoreStatus` in the lib directory
- Update both components to use the new hook
- Use more descriptive variable name `isVectorStoreOrphaned` in GitHubNodeInfo for clarity
